CUHK25CTF-DLLM/Organic_Waste_Organization/server/requirements.txt
Flask==3.1.2
gunicorn==23.0.0
pillow==11.3.0
psycopg2-binary==2.9.10
